G-sensor refactoring. Small UI issues.

+++ SmartRec/SmartRec/GUI/Constants.swift
@@ -117,5 +117,5 @@ let kNavigationBarHeight: CGFloat = 64;
 
 let kLocRadius: Double = 1500.0;
 
-let MAX_ALLOWED_DECELERATION: int = 4; // in g's
+let MAX_ALLOWED_DECELERATION: Double = 4.0; // in g's
